pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "crossref"
version = "0.0.1"
description = "CrossRef: A Python package for automated document cross-referencing"
readme = "README.md"
requires-python = ">=3.11"
license-files = [ "LICENSE" ]
keywords = ["cross-referencing", "document similarity"]
authors = [{ name = "Griffin Holt", email = "griffinbholt@gmail.com" }]
maintainers = [{ name = "Griffin Holt", email = "griffinbholt@gmail.com" }]

# Classifiers help users find your project by categorizing it.
#
# # For a list of valid classifiers, see https://pypi.org/classifiers/
# classifiers = [
#   "Development Status :: 1 - Not Ready",
#   "Intended Audience :: Researchers",
#   "License :: OSI Approved :: MIT License",
#   "Programming Language :: Python :: 3.11",
#   "Programming Language :: Python :: 3.12",
#   "Programming Language :: Python :: 3.13",
#   "Programming Language :: Python :: 3 :: Only",
# ]

# TODO: Determine versions
dependencies = [
    "compose",                # preprocessing.py
    "matplotlib",             # results.py (heatmap)
    "nltk",                   # syntactic.py
    "numpy",                  # base.py, semantic.py
    "pyyaml",                 # main.py
    "sentence_transformers",  # semantic.py
    "torch",                  # semantic.py
]

[project.optional-dependencies]
doc2vec = ["gensim"]                               # Doc2VecMetric, AveragedWordVecMetric, WMDMetric, SoftCosineMetric
tfidf = ["scikit-learn"]                           # TFIDFMetric, LSAMetric
bm25 = ["rank-bm25"]                               # BM25Metric
fuzzy = ["rapidfuzz"]                              # FuzzyNGramMetric
phonetic = ["jellyfish"]                           # PhoneticMetric
colbert = ["pylate"]                               # ColBERTMetric
llm = ["anthropic", "openai", "cohere", "voyageai"]  # LLMJudgeMetric, APIEmbeddingMetric
dev = ["pytest", "coverage"]

# List URLs that are relevant to your project (TODO)
[project.urls]
"Homepage" = "https://github.com/griffinbholt/crossref"
"Bug Reports" = "https://github.com/griffinbholt/crossref/issues"
"Source" = "https://github.com/griffinbholt/crossreft/"

[project.scripts]
crossref = "crossref:main"

[tool.setuptools]
packages = { find = { include = ["crossref*"] } }